+++ lib/password_blacklist.rb
@@ -1,10 +1,10 @@
 require 'password_blacklist/version'
-require 'password_blacklist/tester'
+require 'password_blacklist/checker'
 
 module PasswordBlacklist
   module_function
 
-  def test(password)
-    Tester.new.test(password)
+  def blacklisted?(password)
+    Checker.new.blacklisted?(password)
   end
 end