+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 Unreleased
 ------
-No changes
+Rename `test` methods to `blacklisted?`.
+Rename `PasswordBlacklist::Tester` to `PasswordBlacklist::Checker`.
+Update README.
+Minor changes to resolve Rubcop violations.
 
-0.1.0 / 2017-01-20
+0.1.0.pre / 2017-01-20
 ------
 First release